Added a way to filter proxies by protocol + fixed PremProxy provider
Proxies can now be filtered with an optional parameter in the RequestProxy constructor.
For this the parsers have to set the protocol for each proxy object they build.
Protocols are stored in the ProxyObject.Protocol enum.
PremProxy: parsing was broken because the port number was 'encrypted' and no longer stored in CSS.
They are now obtained from a javascript file that holds a function to the key-port pairs.